refactor(neo4j-storage): extract write mixin (issue #50 sub 2/3)
Issue #50 (EPIC-08-ST-01), Sub-Slice 2/3 — second cut of the
Read/Write/Search split.

- New: backend/app/storage/neo4j_write.py (513 LOC)
  - Neo4jWriteMixin holds all 11 mutating methods:
    - Graph lifecycle: create_graph, delete_graph, set_ontology
    - Add data: add_text (3-phase orchestrator), _persist_episode
      (phase 3, 163 LOC of Cypher), add_text_batch, wait_for_processing
    - Temporal edges (#10): reinforce_relation, tombstone_relation,
      backfill_temporal_defaults
    - Helper: _evaluate_ontology_mutations (best-effort, swallows
      exceptions; moved with the write path because add_text calls it)
  - Imports services.ingestion_pipeline (phases 1+2) and
    storage.neo4j_mappings (edge_to_dict, sanitize_label)
- Changed: backend/app/storage/neo4j_storage.py 701 -> 228 LOC (-473)
  - class Neo4jStorage(Neo4jReadMixin, Neo4jWriteMixin, GraphStorage)
  - Drops uuid, json, typing.List/Callable, services.ingestion_pipeline
    imports — no longer used here
  - Remaining: __init__/close/health-status/_call_with_retry/search +
    the staticmethod re-exports of _node_to_dict/_edge_to_dict

LOC since #50 start: 1127 -> 228 (-899, -79.8% — four fifths gone).

Acceptance criteria:
- [x] "Mappings sind zentral wiederverwendbar" — sub-slice 1
- [ ] "Schreiblogik, Leselogik und Suche liegen nicht mehr in einer
  Datei" — write + read covered, search follows in sub-slice 3

npm run check: 5/5 green (671 backend unchanged, 40 frontend, build ok).
No new tests added — write-path behaviour is wire-identical to the
inline variant and stays covered by test_neo4j_resilience (6 tests),
test_neo4j_ontology_wiring (8), test_ingestion_pipeline (11),
test_neo4j_filtered_entities (4).

Refs #50
