+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Format angelehnt an [Keep a Changelog](https://keepachangelog.com/de/1.1.0/), Ve
 
 ### Geändert
 
+- **Issue #50 (EPIC-08-ST-01: Neo4jStorage in Read/Write/Search schneiden), Sub-Slice 2/3 — Write-Mixin extrahiert.** Alle 11 mutierenden Methoden aus `neo4j_storage.py` ins neue Modul `backend/app/storage/neo4j_write.py` (513 LOC) gezogen: Graph-Lifecycle (`create_graph`, `delete_graph`, `set_ontology`), Add-Data (`add_text` als 3-Phasen-Orchestrator, `_persist_episode` mit 163 LOC Cypher, `add_text_batch`, `wait_for_processing`), Temporal-Edges Issue #10 (`reinforce_relation`, `tombstone_relation`, `backfill_temporal_defaults`) und der Best-Effort-Helper `_evaluate_ontology_mutations`. `Neo4jStorage` erbt jetzt `(Neo4jReadMixin, Neo4jWriteMixin, GraphStorage)` und enthält nur noch Konstruktor, Health-Status, Retry-Wrapper und (vorerst) `search`. Imports `uuid`, `json`, `typing.List`/`Callable` und `services.ingestion_pipeline.*` entfallen aus `neo4j_storage.py`. **`neo4j_storage.py` schrumpft 701 → 228 LOC** (−67,5 %); seit Issue #50-Start von 1127 auf 228 (**−899, −79,8 %, vier Fünftel weg**). Keine neuen Tests — Verhalten Wire-Identical zur Vorher-Variante, gedeckt durch `test_neo4j_resilience` (6), `test_neo4j_ontology_wiring` (8), `test_ingestion_pipeline` (11), `test_neo4j_filtered_entities` (4). **671 Backend-Tests** (unverändert) + 40 Frontend-Tests grün. Folge-Sub-Slice: Search-Mixin (#50/3) schließt EPIC-08 und v0.9.0. Arbeitsprotokoll: `docu/2026-05-01-issue50-sub2-write-mixin-arbeitsprotokoll.md`.
 - **Issue #50 (EPIC-08-ST-01: Neo4jStorage in Read/Write/Search schneiden), Sub-Slice 1/3 — Mappings + Read-Mixin extrahiert.** Erste Etappe des EPIC-08-Boss-Fights. Zwei neue Module: `backend/app/storage/neo4j_mappings.py` (118 LOC) mit den state-losen Helfern `node_to_dict`, `edge_to_dict` und `sanitize_label` (zentral wiederverwendbar — Akzeptanzkriterium ✓), und `backend/app/storage/neo4j_read.py` (381 LOC) mit `Neo4jReadMixin`, das alle 10 Read-Methoden aufnimmt (`get_ontology`, `get_all_nodes`, `get_node`, `get_node_edges`, `get_nodes_by_label`, `get_filtered_entities_with_edges`, `get_all_edges`, `get_edges_at_round`, `get_graph_info`, `get_graph_data`). Mixin-Pattern statt Komposition — `self._driver`/`self._call_with_retry` bleiben über Python-MRO geteilt, kein Konstruktor-Refactoring nötig. **`neo4j_storage.py` schrumpft 1127 → 701 LOC** (−426, −37,8 %, größter Einzelcut in v0.9.0-Pfad-A); `_node_to_dict`/`_edge_to_dict` als `staticmethod`-Re-Export der Modul-Funktionen erhalten, `_sanitize_label` (mitsamt `_LABEL_SAFE_RE`) nach `neo4j_mappings` umgezogen, weil Read- und Write-Pfad sie beide brauchen (zirkuläre Imports vermieden). 23 neue Tests in `test_neo4j_mappings.py` (3 Klassen): node-Mapping inkl. Internal-Field-Stripping und JSON-Fallback, edge-Mapping mit Temporalfeldern und episode_ids-Skalar-Wrapping, sanitize_label inkl. Backtick-Injection-Neutralization und 50-Zeichen-Limit. Bestehende 18 Neo4j-Tests unverändert grün — Verhalten via Mixin-MRO identisch. **671 Backend-Tests** (+23) + 40 Frontend-Tests grün. Folge-Sub-Slices: Write-Mixin (#50/2) und Search-Mixin (#50/3). Arbeitsprotokoll: `docu/2026-05-01-issue50-sub1-mappings-read-arbeitsprotokoll.md`.
 - **Issue #51 (EPIC-08-ST-02: Ingestion-Pipeline in Schritte zerlegen) abgeschlossen.** Die 197-LOC-monolithische Methode `Neo4jStorage.add_text` ist in drei klar getrennte Phasen zerlegt: **Phase 1 (NER + RE)** und **Phase 2 (Batch-Embedding)** liegen als pure Funktionen `extract_entities_and_relations()` und `embed_entities_and_relations()` im neuen Modul `backend/app/services/ingestion_pipeline.py` (97 LOC); **Phase 3 (Persist)** bleibt storage-nah als private Methode `Neo4jStorage._persist_episode` (kwargs-only, 163 LOC, Cypher und Retry 1:1 portiert). `add_text` ist jetzt ein 47-Zeilen-Orchestrator — **−76 % gegenüber dem 197-LOC-Block** vorher. Domain-Logik (NER, Embedding) wandert nach `services/`, Persistenz bleibt in `storage/` — saubere Schichtentrennung als Vorbereitung für #50. 11 neue Tests in `test_ingestion_pipeline.py` (3 Klassen): NER-Delegation und Schema-Pass-Through; Batch-Embedding mit Empty-Skip, Konkatenations-Reihenfolge, Fact-Fallback, Position-Alignment, Crash → Leere-Vektoren, Nur-Entities/Nur-Relations; End-to-End-Komposition Phase 1 → 2. Bestehende Neo4j-Tests (`test_neo4j_resilience`, `test_neo4j_filtered_entities`, `test_neo4j_ontology_wiring`, 18 Tests) unverändert grün — Verhalten identisch. **648 Backend-Tests** (+11) + 40 Frontend-Tests grün. EPIC-08 zu 50 % (2/4); v0.9.0 zu 92 % (11/12). Verbleibend: nur noch **#50** (Boss-Fight, Read/Write/Search-Split). Arbeitsprotokoll: `docu/2026-05-01-issue51-ingestion-pipeline-arbeitsprotokoll.md`.
 - **Issue #48 (EPIC-07-ST-04: Prompt-Building modularisieren) abgeschlossen — EPIC-07 vollständig (5/5).** Zwölf Prompt-Konstanten aus `backend/app/services/report_agent.py` ins neue Modul `backend/app/services/report_prompts.py` (356 LOC) gezogen, semantisch in vier Cluster gegliedert: **Planning** (`PLAN_SYSTEM_PROMPT_TEMPLATE`, `PLAN_USER_PROMPT_TEMPLATE`), **Sections** (`SECTION_SYSTEM_PROMPT_TEMPLATE`, `SECTION_USER_PROMPT_TEMPLATE`), **Reflection / ReACT-Loop** (`REACT_OBSERVATION_TEMPLATE`, `REACT_INSUFFICIENT_TOOLS_MSG`/`_ALT`, `REACT_TOOL_LIMIT_MSG`, `REACT_UNUSED_TOOLS_HINT`, `REACT_FORCE_FINAL_MSG`) und **Chat** (`CHAT_SYSTEM_PROMPT_TEMPLATE`, `CHAT_OBSERVATION_SUFFIX`). Re-Export im Service-Modul hält alle Konsumstellen (`plan_outline`, `_generate_section`, `_run_react_loop`, `chat_with_report`) unverändert. **`report_agent.py` schrumpft 2474 → 2179 LOC** (−295, −11,9 %); seit v0.9.0-Pfad-A-Beginn von 3184 auf 2179 (**−1005 LOC, −31,6 %, ein Drittel weg**). 47 neue Tests in `test_report_prompts.py`: 12 Existenz/Nicht-Leere (parametrisiert), 12 Platzhalter-Inventur (sichert `str.format`-Verträge), `__all__`-Vollständigkeit, 12 Re-Export-Identitätstests, 5 Semantik-Pinning-Tests (JSON-Outline-Forderung, Markdown-Header-Verbot, Tool-Counter-Anzeige, Report-First-Policy, knapper Suffix), 5 Format-Callability-Tests (fängt unbalancierte `{{`/`}}` zur Test- statt Laufzeit). **637 Backend-Tests** (+47) + 40 Frontend-Tests grün. EPIC-07 abgeschlossen; v0.9.0 zu 83 % (10/12). Verbleibend: nur noch EPIC-08 #50, #51 auf `neo4j_storage.py`. Arbeitsprotokoll: `docu/2026-05-01-issue48-prompt-modularisierung-arbeitsprotokoll.md`.
